+++ app/src/main/java/com/pingbits/itchack/ParseData.java
@@ -3,5 +3,5 @@ package com.pingbits.itchack;
 
 public class ParseData {
     public String action;
-    public int data;
+    public float data;
 }
